Add Archive Banner
The Community Plugin has sadly been retired and removed from GitHub. Along with it goes a great source of knowledge and inspiration for many members of this great community. Its an unfortunate end that has happened all too soon.

I am grateful for the opportunity to have played a small part in its existence and wish that things did not need to come to such an unfortunate and premature end. Nonetheless, we move forward like we've always done and will continue building a great community in spite of the challenges. :fist_raised: